pyproject.toml
[tool.poetry]
name = "hierarchy"
version = "0.1.0"
description = ""
authors = ["Trey Dickerson <treydickerson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
